requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bson==0.5.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
defusedxml==0.7.1
fastapi==0.118.0
h11==0.16.0
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
youtube-transcript-api==1.2.3
